backend/requirements.txt
flask
matplotlib
pandas
flask-cors
python-docx
python-dotenv